refactor DateTimeWidgetUtils and DateTimeUtils
