+++ collect_app/src/test/java/org/odk/collect/android/widgets/viewmodels/DateTimeViewModelTest.java
@@ -20,7 +20,6 @@ import org.robolectric.RobolectricTestRunner;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(RobolectricTestRunner.class)
@@ -61,23 +60,20 @@ public class DateTimeViewModelTest {
     }
 
     @Test
-    public void updatingDateInDateSetListener_clearsFocusAndUpdatesDateReturnedByViewModel() {
-        DatePicker datePicker = mock(DatePicker.class);
-        viewModel.getOnDateSetListener().onDateSet(datePicker, 2012, 4, 12);
+    public void updatingDateInDateSetListener_updatesDateReturnedByViewModel() {
+        viewModel.getOnDateSetListener().onDateSet(mock(DatePicker.class), 2012, 4, 12);
 
-        verify(datePicker).clearFocus();
         assertThat(selectedDate.getValue(), is(DateTimeUtils.getSelectedDate(
                 new LocalDateTime().withDate(2012, 5, 12), LocalDateTime.now())));
     }
 
     @Test
-    public void updatingTimeInTimeSetListener_clearsFocusAndUpdatesTimeReturnedByViewModel() {
+    public void updatingTimeInTimeSetListener_updatesTimeReturnedByViewModel() {
         TimePicker timePicker = mock(TimePicker.class);
         when(timePicker.getCurrentHour()).thenReturn(12);
         when(timePicker.getCurrentMinute()).thenReturn(10);
         viewModel.getOnTimeSetListener().onTimeSet(timePicker, 0, 0);
 
-        verify(timePicker).clearFocus();
         assertThat(selectedTime.getValue(), is(DateTimeUtils.getSelectedTime(
                 new LocalDateTime().withTime(12, 10, 0, 0), LocalDateTime.now()).toDateTime()));
     }
